# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AuthForge API
  description: |
    Enterprise Authentication & Authorization Platform API
    
    ## Authentication
    Most endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - Login: 5 attempts per 15 minutes
    - Signup: 3 attempts per hour
    - API: 100 requests per minute (authenticated)
    
    ## Error Codes
    | Code | Description |
    |------|-------------|
    | AUTH_INVALID_CREDENTIALS | Wrong email or password |
    | AUTH_TOKEN_EXPIRED | Access token expired |
    | AUTH_ACCOUNT_LOCKED | Account locked |
    | RBAC_PERMISSION_DENIED | Insufficient permissions |
    
  version: 1.0.0
  contact:
    name: AuthForge Support
    email: support@authforge.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Uses refresh token from HttpOnly cookie
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TokenResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            access_token:
              type: string
            token_type:
              type: string
            expires_in:
              type: integer
...